media-source/SourceBuffer-appendWindowEnd_getting.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Getting SourceBuffer#appendWindowEnd</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

async_test(function(t) {
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e) {
        var sourceBuffer = mediaSource.addSourceBuffer(mime);
        t.step(function() {
            assert_equals(sourceBuffer.appendWindowEnd, Number.POSITIVE_INFINITY, 'Check initial value');
            sourceBuffer.appendWindowEnd = 1;
            assert_equals(sourceBuffer.appendWindowEnd, 1);
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
},
'SourceBuffer#appendWindowEnd : Should return the initial value with positive infinity or the last value that was successfully set');
</script>
</body>
</html>
